refactor(global): 패키지 구조 변경 및 정리
- `config` 패키지를 `global`로 변경하여 일관성 유지
- 관련 클래스들의 패키지 경로를 수정하여 가독성 향상
